Add MPI communication of face data
Finished fixing all the (many) things wrong with MemoryNetwork's
communicate_face_solution. One thing to note is that MPI is not
CUDA-aware on Yellowstone, and this will probably require a custom
install of MPICH to do. So for now, the data is sent to host first, then
MPI-transfered, then sent back to device.